requirements.txt
matplotlib==3.7.1
numpy==1.24.3
pandas==1.5.3
pingouin==0.5.4
Requests==2.32.2
scikit_learn==1.3.0
scipy==1.11.3
seaborn==0.13.2
statsmodels==0.14.0
streamlit==1.33.0
vllm==0.4.2
openai==1.30.1
python-dotenv==1.0.0